+++ contracts/src/token/erc721/extensions/pausable.rs
@@ -5,7 +5,7 @@ use openzeppelin_stylus_proc::r#override;
 use stylus_sdk::{alloy_primitives::U256, msg, prelude::*};
 
 use crate::{
-    token::erc721::{base::IErc721Virtual, Error, TopLevelStorage},
+    token::erc721::{Error, IErc721Virtual, TopLevelStorage},
     utils::Pausable,
 };
 
@@ -38,13 +38,56 @@ impl IErc721Virtual for Erc721PausableOverride {
 #[cfg(all(test, feature = "std"))]
 pub(crate) mod tests {
     use alloy_primitives::address;
+    use stylus_sdk::storage::{StorageBool, StorageMap};
 
     use super::*;
     use crate::token::erc721::{
-        base::Erc721,
-        tests::{random_token_id, Override, Token},
-        traits::IErc721,
+        tests::random_token_id, Erc721, Erc721Override, IErc721,
     };
+
+    sol_storage! {
+        pub struct Token {
+            Erc721<Override> erc721;
+            Erc721Pausable<Override> pausable;
+        }
+    }
+
+    #[r#override]
+    #[inherit(Erc721PausableOverride)]
+    #[inherit(Erc721Override)]
+    impl IErc721Virtual for TokenOverride {}
+
+    unsafe impl TopLevelStorage for Token {}
+
+    impl Default for Token {
+        fn default() -> Self {
+            let root = U256::ZERO;
+
+            Token {
+                erc721: Erc721 {
+                    _owners: unsafe { StorageMap::new(root, 0) },
+                    _balances: unsafe {
+                        StorageMap::new(root + U256::from(32), 0)
+                    },
+                    _token_approvals: unsafe {
+                        StorageMap::new(root + U256::from(64), 0)
+                    },
+                    _operator_approvals: unsafe {
+                        StorageMap::new(root + U256::from(96), 0)
+                    },
+                    _phantom_data: PhantomData,
+                },
+                pausable: Erc721Pausable {
+                    pausable: Pausable {
+                        _paused: unsafe {
+                            StorageBool::new(root + U256::from(128), 0)
+                        },
+                    },
+                    _phantom_data: PhantomData,
+                },
+            }
+        }
+    }
 
     #[motsu::test]
     fn error_transfer_while_paused(contract: Token) {
